Fix several bugs.
1. In console.c, fill the last line with ' ' when scrolling.
2. In loader.c, if start from a direct block, continue loading from
indirect blocks. This is the cause of the mess in memory data when printk and
string.h functions are called.
3. In bootmem.c, fix the physical and virtual address converting
problems. This is the cause of the allocation failure when a multiple of
page size is required. Bitmap operations are changed to function to
avoid potential problems caused by macro argument expansion.
4. Thank gdb, qemu, objdump, readelf.